Improve abstraction and OOP
