+++ frontend/src/components/SignupPage.css
@@ -1,0 +1,5 @@
+.signUpWrapper {
+    /* background-color: blue; */
+    width: 40%;
+    border-radius: 24px;
+}
